Basic node file server, add partials, HTML prettify, build directory.
This is huge refactor. Simply stated, we now have a simple, node-based web server, which removes the need for Pow/Anvil or  MAMP to preview the site builds. We also have html partial files for repeated code (headers, footers), image optimization, and we're prettifying HTML on output. Everything now goes in a build directory and that is what we will push to gh-pages. Lastly, we're managing most of our front-end assets with Bower now, so you should look to Bower first before manually downloading new front-end tools.