fix: 🐛 Remove justified spacing for main nav
The justified spacing for main nav was causing multi-word items to drop to two lines and look hideous. This fix restores left-alignment of the main nav items, and increases the gap between items. The increased spacing is necessary because nav items are uppercase and can be hard to read if they are not spearated by enough space to create clear visual word boundaries. Long nav lists will now break to two lines, which is not ideal, but keeps the nav readable.